Moved the application_group specific functions from schema_utils to resource_mapper. Fixed docs where it says instance instead of group.
