+++ GNUmakefile
@@ -10,7 +10,8 @@ export $(shell sed 's/=.*//' .env)
 # Include .envrc file if it exists
 ifneq (,$(wildcard .envrc))
   include .envrc
-  export $(shell grep -v '^#' .envrc | sed 's/export //')
+  # Commented out problematic line
+  # export $(shell grep -v '^#' .envrc | sed "s/export //")
 endif
 
 default: testacc
