+++ templates/resources/console_application_group_v1.md.tmpl
@@ -13,10 +13,10 @@ This resource allows you to create, read, update and delete application groups i
 
 ## Example Usage
 
-### Simple application instance
+### Simple application group
 {{tffile "examples/resources/conduktor_console_application_group_v1/simple.tf"}}
 
-### Complex application instances
+### Complex application group
 {{tffile "examples/resources/conduktor_console_application_group_v1/complex.tf"}}
